feat(core): unify tilelayer in globeViewer, planarViewer and panoramaViewer
