+++ examples/globe_vector_tiles.html
@@ -31,7 +31,7 @@
             var view = new itowns.GlobeView(viewerDiv, positionOnGlobe, { maxSubdivisionLevel: 13 });
 
             // define pole texture
-            view.wgs84TileLayer.noTextureColor = new itowns.THREE.Color(0x95c1e1);
+            view.tileLayer.noTextureColor = new itowns.THREE.Color(0x95c1e1);
 
             view.atmosphere.visible = false;
