+++ examples/globe_wfs_color.html
@@ -90,7 +90,7 @@
                 console.info('Globe initialized');
                 itowns.ColorLayersOrdering.moveLayerToIndex(globeView, 'Ortho', 0);
             });
-            debug.createTileDebugUI(menuGlobe.gui, globeView, globeView.wgs84TileLayer, d);
+            debug.createTileDebugUI(menuGlobe.gui, globeView, globeView.tileLayer, d);
 </script>
     </body>
 </html>
