+++ examples/oriented_images.html
@@ -150,7 +150,7 @@
             setupPictureUI(menuGlobe, pictureInfos, plane, updatePlaneDistance, globeView, 3000, 15000);
 
             var d = new debug.Debug(globeView, menuGlobe.gui);
-            debug.createTileDebugUI(menuGlobe.gui, globeView, globeView.wgs84TileLayer, d);
+            debug.createTileDebugUI(menuGlobe.gui, globeView, globeView.tileLayer, d);
 
         </script>
     </body>
