+++ examples/panorama.html
@@ -71,8 +71,8 @@
                 // Setup debug menu
                 var gui = new dat.GUI();
                 var ddd = new debug.Debug(view, gui);
-                debug.createTileDebugUI(gui, view, view.baseLayer, ddd);
-                gui.add(view.baseLayer.options, 'quality').min(0.1).max(1.0).onChange(
+                debug.createTileDebugUI(gui, view, view.tileLayer, ddd);
+                gui.add(view.tileLayer.options, 'quality').min(0.1).max(1.0).onChange(
                     function () { view.notifyChange(); });
 
                 // Add controls
