+++ examples/syncCameras.html
@@ -149,7 +149,7 @@
                 },
             });
             var d = new debug.Debug(globeView, menuGlobe.gui);
-            debug.createTileDebugUI(menuGlobe.gui, globeView, globeView.wgs84TileLayer, d);
+            debug.createTileDebugUI(menuGlobe.gui, globeView, globeView.tileLayer, d);
         </script>
     </body>
 </html>
