+++ index.html
@@ -69,7 +69,7 @@ and open the template in the editor.
             });
 
             const d = new debug.Debug(globeView, menuGlobe.gui);
-            debug.createTileDebugUI(menuGlobe.gui, globeView, globeView.wgs84TileLayer, d);
+            debug.createTileDebugUI(menuGlobe.gui, globeView, globeView.tileLayer, d);
             window.globeView = globeView;
 </script>
     </body>
