+++ src/Core/View.js
@@ -3,6 +3,7 @@ import * as THREE from 'three';
 import Camera from '../Renderer/Camera';
 import MainLoop, { MAIN_LOOP_EVENTS, RENDERING_PAUSED } from './MainLoop';
 import c3DEngine from '../Renderer/c3DEngine';
+import RendererConstant from '../Renderer/RendererConstant';
 
 import { getMaxColorSamplerUnitsCount } from '../Renderer/LayeredMaterial';
 
@@ -621,6 +622,114 @@ View.prototype.pickObjectsAt = function pickObjectsAt(mouseOrEvt, radius, ...whe
     }
 
     return results;
+};
+
+View.prototype.readDepthBuffer = function readDepthBuffer(x, y, width, height) {
+    const g = this.mainLoop.gfxEngine;
+    const currentWireframe = this.tileLayer.wireframe;
+    const currentOpacity = this.tileLayer.opacity;
+    const currentVisibility = this.tileLayer.visible;
+    if (currentWireframe) {
+        this.tileLayer.wireframe = false;
+    }
+    if (currentOpacity < 1.0) {
+        this.tileLayer.opacity = 1.0;
+    }
+    if (!currentVisibility) {
+        this.tileLayer.visible = true;
+    }
+
+    const restore = this.tileLayer.level0Nodes.map(n => n.pushRenderState(RendererConstant.DEPTH));
+    const buffer = g.renderViewToBuffer(
+        { camera: this.camera, scene: this.tileLayer.object3d },
+        { x, y, width, height });
+    restore.forEach(r => r());
+
+    if (this.tileLayer.wireframe !== currentWireframe) {
+        this.tileLayer.wireframe = currentWireframe;
+    }
+    if (this.tileLayer.opacity !== currentOpacity) {
+        this.tileLayer.opacity = currentOpacity;
+    }
+    if (this.tileLayer.visible !== currentVisibility) {
+        this.tileLayer.visible = currentVisibility;
+    }
+
+    return buffer;
+};
+
+const matrix = new THREE.Matrix4();
+const screen = new THREE.Vector2();
+const ray = new THREE.Ray();
+const direction = new THREE.Vector3();
+
+/**
+ * Returns the world position (view's crs: referenceCrs) under view coordinates.
+ * This position is computed with depth buffer.
+ *
+ * @param      {THREE.Vector2}  mouse  position in view coordinates (in pixel), if it's null so it's view's center.
+ * @param      {THREE.Vector3}  [target=THREE.Vector3()] target. the result will be copied into this Vector3. If not present a new one will be created.
+ * @return     {THREE.Vector3}  the world position in view's crs: referenceCrs.
+ */
+
+View.prototype.getPickingPositionFromDepth = function fnGetPickingPositionFromDepth(mouse, target = new THREE.Vector3()) {
+    if (!this.tileLayer || this.tileLayer.level0Nodes.length == 0 || (!this.tileLayer.level0Nodes[0])) {
+        target = undefined;
+        return;
+    }
+    const l = this.mainLoop;
+    const viewPaused = l.scheduler.commandsWaitingExecutionCount() == 0 && l.renderingState == RENDERING_PAUSED;
+    const g = l.gfxEngine;
+    const dim = g.getWindowSize();
+    const camera = this.camera.camera3D;
+
+    mouse = mouse || dim.clone().multiplyScalar(0.5);
+    mouse.x = Math.floor(mouse.x);
+    mouse.y = Math.floor(mouse.y);
+
+    // Prepare state
+    const prev = camera.layers.mask;
+    camera.layers.mask = 1 << this.tileLayer.threejsLayer;
+
+     // Render/Read to buffer
+    let buffer;
+    if (viewPaused) {
+        this._fullSizeDepthBuffer = this._fullSizeDepthBuffer || this.readDepthBuffer(0, 0, dim.x, dim.y);
+        const id = ((dim.y - mouse.y - 1) * dim.x + mouse.x) * 4;
+        buffer = this._fullSizeDepthBuffer.slice(id, id + 4);
+    } else {
+        buffer = this.readDepthBuffer(mouse.x, mouse.y, 1, 1);
+    }
+
+    screen.x = (mouse.x / dim.x) * 2 - 1;
+    screen.y = -(mouse.y / dim.y) * 2 + 1;
+
+    // Origin
+    ray.origin.copy(camera.position);
+
+    // Direction
+    ray.direction.set(screen.x, screen.y, 0.5);
+    // Unproject
+    matrix.multiplyMatrices(camera.matrixWorld, matrix.getInverse(camera.projectionMatrix));
+    ray.direction.applyMatrix4(matrix);
+    ray.direction.sub(ray.origin);
+
+    direction.set(0, 0, 1.0);
+    direction.applyMatrix4(matrix);
+    direction.sub(ray.origin);
+
+    const angle = direction.angleTo(ray.direction);
+    const orthoZ = g.depthBufferRGBAValueToOrthoZ(buffer, camera);
+    const length = orthoZ / Math.cos(angle);
+
+    target.addVectors(camera.position, ray.direction.setLength(length));
+
+    camera.layers.mask = prev;
+
+    if (target.length() > 10000000)
+        { return undefined; }
+
+    return target;
 };
 
 export default View;
